Clang-tidy

+++ comm/x.cpp
@@ -25,7 +25,7 @@ int x() {
     k = (main_hero->station) ? (0.8) : (0.5);
 
     if (main_hero->get_stuff() > 0) {
-        main_hero->add_money((int) main_hero->get_stuff() * k + 1);
+        main_hero->add_money((int) ((float) main_hero->get_stuff() * k) + 1);
 
         main_hero->sub_stuff(main_hero->get_stuff());
